requirements.txt
langchain
langchain_core
langchain_community
streamlit
openai
python-dotenv
google.genai
